fix: missing icon exports and fix react flow rendering
