+++ app/src/components/visualizer/index.ts
@@ -1,6 +1,7 @@
 export { default as EditorActions } from "./EditorActions/EditorActions";
 export { default as EditorControls } from "./EditorControls/EditorControls";
 export { default as GardenFlow } from "./GardenFlow/GardenFlow";
+export { default as GardenFlowInner } from "./GardenFlowInner/GardenFlowInner";
 export { default as GardenTabs } from "./GardenTabs/GardenTabs";
 export { default as SchemaEditor } from "./SchemaEditor/SchemaEditor";
 export { default as SchemaHelp } from "./SchemaHelp/SchemaHelp";
